Move javascript from index.html to custom.js
We want to execute the "Self Executing Function", located at the bottom
of the body. We can also achieve that with a custom function in the
"onload" attribute of the body tag. This way we keep the HTML as clean
as possible.

The javascript for the enviroment warning still works if it is moved to
the custom initialize function.